Se ha corregido el Bug que impedía usar el generador de assets aleatorios si se usaba /mute respondiendo un mensaje. Además, si se responde a un mensaje y el usuario está silenciado, el bot responderá con un mensaje de error indicando que dicho usuario ya está silenciado

+++ src/commands/admin_commands/mute.rs
@@ -33,12 +33,55 @@ pub async fn mute_user_admin(bot: Bot, msg: Message) -> ResponseResult<()> {
                 let is_admin_or_owner = chat_member.status().is_administrator() || chat_member.status().is_owner();
 
                 if is_admin_or_owner {
-                    bot.restrict_chat_member(msg.chat.id, user.id, ChatPermissions::empty()).await?;
-                    bot.send_message(msg.chat.id, format!("✅ @{} ha sido silenciado", username_user)).await?;
 
-                    sleep(Duration::from_secs(5)).await;
-                    bot.delete_message(msg.chat.id, msg.id).await?;
-                    mute_animation_generator(bot, msg).await?;
+                    let chat_member = bot.get_chat_member(msg.chat.id, user.id).await?;
+                    if chat_member.status().is_restricted() {
+                        let err = bot.send_message(msg.chat.id, format!
+                        ("❌ @{} Ya está silenciado. Usa este comando solo para silenciar a alguien que no esté silenciado", username_user))
+                            .reply_to_message_id(msg.id)
+                            .parse_mode(ParseMode::Html)
+                            .await?;
+
+                        sleep(Duration::from_secs(10)).await;
+                        bot.delete_message(msg.chat.id, err.id).await?;
+                        bot.delete_message(msg.chat.id, msg.id).await?;
+                        return Ok(());
+
+                    } else {
+                        bot.restrict_chat_member(msg.chat.id, user.id, ChatPermissions::empty()).await?;
+
+                        let mut rng: StdRng = SeedableRng::from_entropy();
+                        let random_number = rng.gen_range(0..=14);
+                        let file_names = [
+                            "1.gif", "2.gif", "3.gif", "4.gif", "5.jpg",
+                        ];
+
+                        let get_file_name = |index: usize| -> &'static str {
+                            file_names.get(index).unwrap_or_else(|| file_names.last().unwrap())
+                        };
+                        let file_path = format!("./assets/mute/{}", get_file_name(random_number));
+
+                        match file_path.ends_with(".gif") {
+
+                            true => {
+                                bot.send_animation(msg.chat.id, InputFile::file(file_path))
+                                    .caption(format!("✅ @{} [<code>{}</code>] silenciado", username_user, user.id))
+                                    .parse_mode(ParseMode::Html)
+                                    .reply_to_message_id(msg.id)
+                                    .await?;
+                            }
+
+                            false => {
+                                bot.send_photo(msg.chat.id, InputFile::file(file_path))
+                                    .caption(format!("✅ @{} <code>[{}</code>] silenciado", username_user, user.id))
+                                    .parse_mode(ParseMode::Html)
+                                    .reply_to_message_id(msg.id)
+                                    .await?;
+                            }
+
+                        }
+
+                    }
 
                 } else {
                     let err = bot.send_message(msg.chat.id, "❌ No tienes permisos para silenciar a un usuario").await?;
